requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
face-recognition==1.3.0
numpy==1.24.3
opencv-python==4.8.1.78
cmake==3.27.7
dlib==19.24.0
bcrypt==4.0.1
